Oops init bug

+++ horde-init.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
-cp -rp /etc/.horde/* /etc/horde/
-cp /etc/horde/horde/conf.php.dist /etc/horde/horde/conf.php
+if [[ ! -f "/etc/horde/horde/conf.php" ]]; then
+    cp -rp /etc/.horde/* /etc/horde/
+    cp /etc/horde/horde/conf.php.dist /etc/horde/horde/conf.php
+fi
